refactor(tests): split ghcr_prune_untagged tests + reuse fetched versions
- prune_package gains a `versions` kwarg; main() pre-fetches versions for
  reference-digest collection and reuses them, eliminating a duplicate
  GitHub API call per protected package.

- Split the 1k-line monolithic test file into a focused package under
  tests/scripts/ghcr_prune/:
    _mocks.py         shared HTTP fakes (owner/versions/delete/manifest)
    test_helpers.py   GhcrVersion, list_container_versions, delete_version
    test_main.py      main() integration (untagged-only, keep_n+dry_run)
    test_buildcache.py buildcache retention (pr/mq/dispatch tags) +
                      main iteration over BUILDCACHE_PACKAGES
    test_referenced.py fetch_manifest, collect_referenced_digests,
                       prune_package referenced_digests, end-to-end SLSA
  Parametrized the ephemeral-tag deletion test across pr-/mq-/dispatch-
  to drop three near-duplicate cases.

Refs #884.
