+++ tests/scripts/ghcr_prune/__init__.py
@@ -1,0 +1,1 @@
+"""Test package for ``scripts.ci.maintenance.ghcr_prune_untagged``."""
